pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Web service which provides the oppotunity for task management, including adding tasks with different pamameters referring to the users of the app, their tracking and filtering."
authors = ["Dddarknight <9801677@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"
keywords = ["task manager", "django", "json", "yaml", "yml"]
classifiers = [
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Programming Language :: Python :: 3",
  "Framework :: Django :: 4.1",
  "Topic :: Internet :: WWW/HTTP :: WSGI :: Application"
]
repository = "https://github.com/Dddarknight/python-project-52"
packages = [
  { include = "task_manager" },
]

[tool.poetry.dependencies]
python = "^3.8"
dj-database-url = ">=0.5.0,<=1.0.0"
Django = "^4.1"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"
django-bootstrap4 = "^22.2"
django-filter = "^22.1"
rollbar = "^0.16.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.6.0"
flake8 = "^5.0.4"
pytest = "^7.1.2"
coverage = "^6.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
